Replaced ndomod.cfg data_processing_options variable with single options.
Previous versions of ndoutils required the use of a bitmask to enable/disable which data sets were sent to the database. These are now defined in ndomod.cfg as single booleans, which are more human readable. Example:

acknowledgement_data=1
adaptive_contact_data=1

This should allow easier customization, and allow users to reduce overhead for data that they don't need. 

Updated README and CHANGELOG as well.