Modified behavior of upgradedb script to support multiple updates.
This commit modifies the meaning of the of the dbversion table and
the behavior of the db/upgradedb script in the following ways:

1. The dbversion is now redefined semantically to mean the schema version
   of the database. Previously the version was the same as the version of
   the NDOUtils code which meant there could be an update to the version in
   the dbversion table when there was really no schema change.
2. The db/upgradedb script has been modified to iteratively apply all
   mysql-upgrade-<version>.sql scripts between the current version as stored
   in the dbversion table and the latest version supported by the script.
   See the comment in the db/upgradedb script immediately above the
   definition of the @schemaversions variable for more information.

This commit also makes possible the implementation of a target version other
than the latest version supported by the upgradedb script. It also makes
possible downgrading versions of the schema. These features have *not* been
implement yet, but they may be in future versions.