ndomod: broke ndomod_broker_data into bite-sized pieces
This commit breakes the ndomod_broker_data() function in smaller pieces,
one for each event type. (Previously it was 2k lines). The event-specific
functions are called though a function pointer stored in a table with one
entry for each function. The event-specific functions can also perform
different actions based on the phase passed to them, enabling
ndomod_broker_data() to still behave as it always has.