nod2db: Message queue send and retry timing fixes.
Instead of a non-blocking msgsnd() followed by sleep() and a retry on
failure, set an alarm() on a blocking msgsnd(). A one-second alarm lets
us log retries similarly to what was done before: the difference being
sends that don't complete immediately but do complete within one second
are no longer retires. In testing, the vast majority of reties were
completed within one second and are no longer logged.

The alarm() signal handling should not interfere with
ndo2db_child_sighandler(), as that simply calls _exit(0). Signals
that don't result in process termination but do EINTR the msgsnd() are
seen as an alarm expiration, so this scheme is not perfect..