From PR https://github.com/NagiosEnterprises/ndoutils/pull/3 (David Michael)
Descriptions from his original commits below

ndo2db: Have the parent process wait until the lock file is written.

When a service has type "forking", systemd tries to read its PID
file after the parent process exits.  This change avoids a race
condition where the parent process can sometimes exit before the
PID is written to the lock file by the grandchild process.

ndo2db: Accept a socket fd passed from systemd.

This defines a configure test that adds the systemd library to LIBS
if it could be found.  It also adds an option to explicitly enable
or disable the configure test, so e.g. it can return an error if
systemd was requested and not found.