Support long names in nagios_objects.
It seems like we chose short varchar lengths in nagios_objects to increase speed
(using the `uniq_object` key to index the table). 128 or 162 are the shortest 3-byte
unicode string lengths we can have in the key before older mysql versions complain.

However, forcing short names in the table will cause collisions, due to the unlimited name length
allowed in Nagios Core. This can cause e.g. check results to be shared between multiple objects
within the database.

To get around this, we're using longer column lengths and instead using prefixes of the name1 and name2
columns within the key itself. The application doesn't seem to rely on the key being unique, and we're
expecting that there will be collisions now, so the unique key is degraded to a regular key