src/Makefile.in: install all executables with mode 0755
Three executables -- file2sock, log2ndo, and sockdebug -- are
currently being installed group-writable but not
world-executable. This is in contrast with the other two executables,
ndo2db and ndomod.o, that are installed mode 0755.

Having recently removed the INSTALL_OPTS that were altering the
owner:group of these files, there is no longer any security risk to
mode 0774. However, 0755 is more consistent with both the rest of our
executables, and with the typical permissions on /usr/bin that arise
from the (extremely common) umask of 0022.

We change these three to 0755 for a little bit of extra peace of mind.

changes. Lines starting # with '#' will be ignored, and an empty
message aborts the commit.  # # Date: Sat Mar 2 19:52:47 2024 -0500 #
src/Makefile.in #